backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-storages==1.14.4

psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0

gunicorn==23.0.0
whitenoise==6.8.2

# Payments
stripe==11.3.0

# Google Calendar Integration
google-api-python-client==2.155.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Microsoft Outlook Integration
msal==1.31.0
requests-oauthlib==2.0.0

# Timezone
pytz==2024.2
python-dateutil==2.9.0.post0

# Utilities
Pillow==11.0.0
python-slugify==8.0.4
uuid==1.30
cryptography==44.0.0

# Email
premailer==3.10.0

# API documentation
drf-spectacular==0.28.0

# Testing
factory-boy==3.3.1
faker==33.0.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
